requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.7
django-admin-interface==0.20.0
django-celery-beat==2.3.0
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-jet-reboot==1.3.1
django-material-admin==1.8.6
django-messages==0.6.0
django-timezone-field==5.0
feedparser==6.0.10
google-api-python-client==1.4.1
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
kombu==5.2.4
oauth2client==4.1.3
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
rsa==4.9
selenium==4.4.3
sgmllib3k==1.0.0
shortuuid==1.0.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
